+++ src/application/battlespace/BattleSpaceGame.cpp
@@ -40,43 +40,43 @@ void BattleSpaceGame::initialize() {
     setGlRenderStyle(LINES);
     setGlShadeModel(FLAT);
 
-    addEntity(new Astronomical("planet1", Util::Math::Vector3<double>(13, -15, -1000), 150, Util::Math::Vector3<double>(1, 0, 0), Util::Graphic::Colors::BLUE));
-    addEntity(new Astronomical("planet1", Util::Math::Vector3<double>(37, -8, 3000), 400, Util::Math::Vector3<double>(0.25, 0, 0), Util::Graphic::Colors::YELLOW));
+    addEntity(new Astronomical("planet1", Util::Math::Vector3<float>(13, -15, -1000), 150, Util::Math::Vector3<float>(1, 0, 0), Util::Graphic::Colors::BLUE));
+    addEntity(new Astronomical("planet1", Util::Math::Vector3<float>(37, -8, 3000), 400, Util::Math::Vector3<float>(0.25, 0, 0), Util::Graphic::Colors::YELLOW));
 
     auto modelId = static_cast<uint8_t>(random.getRandomNumber() * 2 + 1);
     auto scale = 5 + random.getRandomNumber() * 10;
-    auto rotation = Util::Math::Vector3<double>(random.getRandomNumber() * 5, random.getRandomNumber() * 5, random.getRandomNumber() * 5);
-    addEntity(new Astronomical(Util::String::format("asteroid%u", modelId), Util::Math::Vector3<double>(-25, 12, -42), scale, rotation, Util::Graphic::Colors::BROWN));
+    auto rotation = Util::Math::Vector3<float>(random.getRandomNumber() * 5, random.getRandomNumber() * 5, random.getRandomNumber() * 5);
+    addEntity(new Astronomical(Util::String::format("asteroid%u", modelId), Util::Math::Vector3<float>(-25, 12, -42), scale, rotation, Util::Graphic::Colors::BROWN));
 
     modelId = static_cast<uint8_t>(random.getRandomNumber() * 2 + 1);
     scale = 5 + random.getRandomNumber() * 10;
-    rotation = Util::Math::Vector3<double>(random.getRandomNumber() * 5, random.getRandomNumber() * 5, random.getRandomNumber() * 5);
-    addEntity(new Astronomical(Util::String::format("asteroid%u", modelId), Util::Math::Vector3<double>(40, 30, -55), scale, rotation, Util::Graphic::Colors::BROWN));
+    rotation = Util::Math::Vector3<float>(random.getRandomNumber() * 5, random.getRandomNumber() * 5, random.getRandomNumber() * 5);
+    addEntity(new Astronomical(Util::String::format("asteroid%u", modelId), Util::Math::Vector3<float>(40, 30, -55), scale, rotation, Util::Graphic::Colors::BROWN));
 
     modelId = static_cast<uint8_t>(random.getRandomNumber() * 2 + 1);
     scale = 5 + random.getRandomNumber() * 10;
-    rotation = Util::Math::Vector3<double>(random.getRandomNumber() * 5, random.getRandomNumber() * 5, random.getRandomNumber() * 5);
-    addEntity(new Astronomical(Util::String::format("asteroid%u", modelId), Util::Math::Vector3<double>(78, -24, 63), scale, rotation, Util::Graphic::Colors::BROWN));
+    rotation = Util::Math::Vector3<float>(random.getRandomNumber() * 5, random.getRandomNumber() * 5, random.getRandomNumber() * 5);
+    addEntity(new Astronomical(Util::String::format("asteroid%u", modelId), Util::Math::Vector3<float>(78, -24, 63), scale, rotation, Util::Graphic::Colors::BROWN));
 
     modelId = static_cast<uint8_t>(random.getRandomNumber() * 2 + 1);
     scale = 5 + random.getRandomNumber() * 10;
-    rotation = Util::Math::Vector3<double>(random.getRandomNumber() * 5, random.getRandomNumber() * 5, random.getRandomNumber() * 5);
-    addEntity(new Astronomical(Util::String::format("asteroid%u", modelId), Util::Math::Vector3<double>(-92, -74, 48), scale, rotation, Util::Graphic::Colors::BROWN));
+    rotation = Util::Math::Vector3<float>(random.getRandomNumber() * 5, random.getRandomNumber() * 5, random.getRandomNumber() * 5);
+    addEntity(new Astronomical(Util::String::format("asteroid%u", modelId), Util::Math::Vector3<float>(-92, -74, 48), scale, rotation, Util::Graphic::Colors::BROWN));
 
     modelId = static_cast<uint8_t>(random.getRandomNumber() * 2 + 1);
     scale = 5 + random.getRandomNumber() * 10;
-    rotation = Util::Math::Vector3<double>(random.getRandomNumber() * 5, random.getRandomNumber() * 5, random.getRandomNumber() * 5);
-    addEntity(new Astronomical(Util::String::format("asteroid%u", modelId), Util::Math::Vector3<double>(-48, 30, 37), scale, rotation, Util::Graphic::Colors::BROWN));
+    rotation = Util::Math::Vector3<float>(random.getRandomNumber() * 5, random.getRandomNumber() * 5, random.getRandomNumber() * 5);
+    addEntity(new Astronomical(Util::String::format("asteroid%u", modelId), Util::Math::Vector3<float>(-48, 30, 37), scale, rotation, Util::Graphic::Colors::BROWN));
 
     modelId = static_cast<uint8_t>(random.getRandomNumber() * 2 + 1);
     scale = 5 + random.getRandomNumber() * 10;
-    rotation = Util::Math::Vector3<double>(random.getRandomNumber() * 5, random.getRandomNumber() * 5, random.getRandomNumber() * 5);
-    addEntity(new Astronomical(Util::String::format("asteroid%u", modelId), Util::Math::Vector3<double>(10, 23, 78), scale, rotation, Util::Graphic::Colors::BROWN));
+    rotation = Util::Math::Vector3<float>(random.getRandomNumber() * 5, random.getRandomNumber() * 5, random.getRandomNumber() * 5);
+    addEntity(new Astronomical(Util::String::format("asteroid%u", modelId), Util::Math::Vector3<float>(10, 23, 78), scale, rotation, Util::Graphic::Colors::BROWN));
 
     addEntity(player);
 }
 
-void BattleSpaceGame::update(double delta) {
+void BattleSpaceGame::update(float delta) {
     if (player->getHealth() <= 0) {
         Pulsar::Game::getInstance().pushScene(new Pulsar::TextScreen(Util::String::format(GAME_OVER_TEXT, player->getScore()), handleKeyPressOnTextScreen, Util::Graphic::Colors::GREEN));
         Pulsar::Game::getInstance().switchToNextScene();
@@ -96,7 +96,7 @@ void BattleSpaceGame::update(double delta) {
         }
 
         auto &camera = getCamera();
-        camera.setRotation(Util::Math::Vector3<double>(newRotation.getX(), newPitch, newRotation.getZ()));
+        camera.setRotation(Util::Math::Vector3<float>(newRotation.getX(), newPitch, newRotation.getZ()));
 
         auto translation = inputTranslation + camera.getFrontVector() * player->getSpeed();
         camera.translate(translation * delta);
@@ -107,8 +107,8 @@ void BattleSpaceGame::update(double delta) {
         // Spawn enemies
         if (enemies.isEmpty()) {
             difficulty++;
-            double minEnemies = (difficulty / 3.0);
-            double enemyCount =  minEnemies + random.getRandomNumber() * (difficulty - minEnemies);
+            float minEnemies = (difficulty / 3.0);
+            float enemyCount =  minEnemies + random.getRandomNumber() * (difficulty - minEnemies);
 
             for (uint32_t i = 0; i < enemyCount; i++) {
                 auto x = (random.getRandomNumber() * (ENEMY_SPAWN_RANGE * 2)) - ENEMY_SPAWN_RANGE / 2.0;
@@ -116,7 +116,7 @@ void BattleSpaceGame::update(double delta) {
                 auto z = (random.getRandomNumber() * (ENEMY_SPAWN_RANGE * 2)) - ENEMY_SPAWN_RANGE / 2.0;
                 auto type = random.getRandomNumber() * 5;
 
-                auto *enemy = new Enemy(*player, enemies, Util::Math::Vector3<double>(x, y / 2, z), Util::Math::Vector3<double>(0, 0, 0), 1, static_cast<Enemy::Type>(type));
+                auto *enemy = new Enemy(*player, enemies, Util::Math::Vector3<float>(x, y / 2, z), Util::Math::Vector3<float>(0, 0, 0), 1, static_cast<Enemy::Type>(type));
                 enemies.add(enemy);
                 addEntity(enemy);
             }
@@ -130,28 +130,28 @@ void BattleSpaceGame::keyPressed(const Util::Io::Key &key) {
             Pulsar::Game::getInstance().stop();
             break;
         case Util::Io::Key::LEFT:
-            inputRotation = Util::Math::Vector3<double>(inputRotation.getX(), inputRotation.getY(), -1);
+            inputRotation = Util::Math::Vector3<float>(inputRotation.getX(), inputRotation.getY(), -1);
         break;
         case Util::Io::Key::RIGHT:
-            inputRotation = Util::Math::Vector3<double>(inputRotation.getX(), inputRotation.getY(), 1);
+            inputRotation = Util::Math::Vector3<float>(inputRotation.getX(), inputRotation.getY(), 1);
         break;
         case Util::Io::Key::UP:
-            inputRotation = Util::Math::Vector3<double>(inputRotation.getX(), 1, inputRotation.getZ());
+            inputRotation = Util::Math::Vector3<float>(inputRotation.getX(), 1, inputRotation.getZ());
         break;
         case Util::Io::Key::DOWN:
-            inputRotation = Util::Math::Vector3<double>(inputRotation.getX(), -1, inputRotation.getZ());
+            inputRotation = Util::Math::Vector3<float>(inputRotation.getX(), -1, inputRotation.getZ());
         break;
         case Util::Io::Key::W:
-            inputTranslation = Util::Math::Vector3<double>(inputTranslation.getX(), 1, inputTranslation.getZ());
+            inputTranslation = Util::Math::Vector3<float>(inputTranslation.getX(), 1, inputTranslation.getZ());
         break;
         case Util::Io::Key::S:
-            inputTranslation = Util::Math::Vector3<double>(inputTranslation.getX(), -1, inputTranslation.getZ());
+            inputTranslation = Util::Math::Vector3<float>(inputTranslation.getX(), -1, inputTranslation.getZ());
         break;
         case Util::Io::Key::A:
-            inputTranslation = Util::Math::Vector3<double>(-1, inputTranslation.getY(), inputTranslation.getZ());
+            inputTranslation = Util::Math::Vector3<float>(-1, inputTranslation.getY(), inputTranslation.getZ());
         break;
         case Util::Io::Key::D:
-            inputTranslation = Util::Math::Vector3<double>(1, inputTranslation.getY(), inputTranslation.getZ());
+            inputTranslation = Util::Math::Vector3<float>(1, inputTranslation.getY(), inputTranslation.getZ());
         break;
         case Util::Io::Key::Q:
             inputSpeed -= 0.1;
@@ -170,7 +170,7 @@ void BattleSpaceGame::keyPressed(const Util::Io::Key &key) {
 
         case Util::Io::Key::ENTER: {
             const auto &camera = getCamera();
-            auto *enemy = new Enemy(*player, enemies, camera.getPosition() + camera.getFrontVector() * 10, Util::Math::Vector3<double>(0, 0, 0), 1, Enemy::Type::STATIONARY);
+            auto *enemy = new Enemy(*player, enemies, camera.getPosition() + camera.getFrontVector() * 10, Util::Math::Vector3<float>(0, 0, 0), 1, Enemy::Type::STATIONARY);
             enemies.add(enemy);
             addEntity(enemy);
         }
@@ -183,19 +183,19 @@ void BattleSpaceGame::keyReleased(const Util::Io::Key &key) {
     switch (key.getScancode()) {
         case Util::Io::Key::LEFT:
         case Util::Io::Key::RIGHT:
-            inputRotation = Util::Math::Vector3<double>(inputRotation.getX(), inputRotation.getY(), 0);
+            inputRotation = Util::Math::Vector3<float>(inputRotation.getX(), inputRotation.getY(), 0);
         break;
         case Util::Io::Key::UP:
         case Util::Io::Key::DOWN:
-            inputRotation = Util::Math::Vector3<double>(inputRotation.getX(), 0, inputRotation.getZ());
+            inputRotation = Util::Math::Vector3<float>(inputRotation.getX(), 0, inputRotation.getZ());
         break;
         case Util::Io::Key::W:
         case Util::Io::Key::S:
-            inputTranslation = Util::Math::Vector3<double>(inputTranslation.getX(), 0, inputTranslation.getZ());
+            inputTranslation = Util::Math::Vector3<float>(inputTranslation.getX(), 0, inputTranslation.getZ());
         break;
         case Util::Io::Key::A:
         case Util::Io::Key::D:
-            inputTranslation = Util::Math::Vector3<double>(0, inputTranslation.getY(), inputTranslation.getZ());
+            inputTranslation = Util::Math::Vector3<float>(0, inputTranslation.getY(), inputTranslation.getZ());
         break;
         default:
             break;
