+++ src/application/battlespace/EnemyDebris.cpp
@@ -29,7 +29,7 @@
 #include "lib/util/graphic/Colors.h"
 #include "lib/util/math/Random.h"
 
-EnemyDebris::EnemyDebris(const Util::Math::Vector3<double> &position, const Util::Math::Vector3<double> &rotation, double scale, uint8_t modelId) : Pulsar::D3::Model(Enemy::TAG, Util::String::format("/user/battlespace/debris%u.obj", modelId), position, rotation, Util::Math::Vector3<double>(scale, scale, scale), Util::Graphic::Colors::RED), modelId(modelId) {}
+EnemyDebris::EnemyDebris(const Util::Math::Vector3<float> &position, const Util::Math::Vector3<float> &rotation, float scale, uint8_t modelId) : Pulsar::D3::Model(Enemy::TAG, Util::String::format("/user/battlespace/debris%u.obj", modelId), position, rotation, Util::Math::Vector3<float>(scale, scale, scale), Util::Graphic::Colors::RED), modelId(modelId) {}
 
 void EnemyDebris::initialize() {
     Model::initialize();
@@ -38,19 +38,19 @@ void EnemyDebris::initialize() {
         case 1: {
             auto r = Util::Math::Random().getRandomNumber();
             auto r2 = (r - 0.5) * 2.0;
-            translateDirection = Util::Math::Vector3<double>(-0.05 - 0.1 * r, 0.05 * r2, 0.05 + 0.05 * r);
+            translateDirection = Util::Math::Vector3<float>(-0.05 - 0.1 * r, 0.05 * r2, 0.05 + 0.05 * r);
             break;
         }
         case 2: {
             auto r = Util::Math::Random().getRandomNumber();
             auto r2 = (r - 0.5) * 2.0;
-            translateDirection = Util::Math::Vector3<double>(0.05 + 0.1 * r, 0.05 * r2, 0.05 + 0.05 * r);
+            translateDirection = Util::Math::Vector3<float>(0.05 + 0.1 * r, 0.05 * r2, 0.05 + 0.05 * r);
             break;
         }
         case 3: {
             auto r = Util::Math::Random().getRandomNumber();
             auto r2 = (r - 0.5) * 2.0;
-            translateDirection = Util::Math::Vector3<double>(-0.05 - 0.1 * r, 0.05 * r2, 0.05 + 0.05 * r);
+            translateDirection = Util::Math::Vector3<float>(-0.05 - 0.1 * r, 0.05 * r2, 0.05 + 0.05 * r);
             break;
         }
         default:
@@ -58,10 +58,10 @@ void EnemyDebris::initialize() {
     }
 
     auto r = Util::Math::Random().getRandomNumber() * 35;;
-    rotationDirection = Util::Math::Vector3<double>(r, r, r);
+    rotationDirection = Util::Math::Vector3<float>(r, r, r);
 }
 
-void EnemyDebris::onUpdate(double delta) {
+void EnemyDebris::onUpdate(float delta) {
     lifetime += delta;
     if (lifetime > 2) {
         removeFromScene();
