+++ src/application/bug/EnemyBug.h
@@ -39,7 +39,7 @@ public:
     /**
      * Constructor.
      */
-    explicit EnemyBug(const Util::Math::Vector2<double> &position, Fleet &fleet);
+    explicit EnemyBug(const Util::Math::Vector2<float> &position, Fleet &fleet);
 
     /**
      * Copy Constructor.
@@ -58,7 +58,7 @@ public:
 
     void initialize() override;
 
-    void onUpdate(double delta) override;
+    void onUpdate(float delta) override;
 
     void onTranslationEvent(Pulsar::D2::TranslationEvent &event) override;
 
@@ -69,15 +69,15 @@ public:
     void fireMissile();
 
     static const constexpr uint32_t TAG = 3;
-    static const constexpr double SIZE_X = 0.15;
-    static const constexpr double SIZE_Y = 0.1;
+    static const constexpr float SIZE_X = 0.15;
+    static const constexpr float SIZE_Y = 0.1;
 
 private:
 
     Pulsar::D2::SpriteAnimation animation;
     Fleet &fleet;
 
-    double lastMissileRollTime = 0;
+    float lastMissileRollTime = 0;
 };
 
 #endif
